requirements.txt
asn1crypto==1.2.0
beautifulsoup4==4.8.1
boto==2.49.0
boto3==1.9.234
botocore==1.12.234
bz2file==0.98
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
cssselect==1.1.0
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
gensim==3.8.0
idna==2.8
jieba3k==0.35.1
jmespath==0.9.4
lxml==4.4.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.2
Pillow==6.2.1
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
smart-open==1.8.4
soupsieve==1.9.4
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.6
tabulate
